Move some files / config around
