+++ src/server.js
@@ -1,7 +1,7 @@
 var webpack = require('webpack');
 var webpackDevMiddleware = require('webpack-dev-middleware');
 var webpackHotMiddleware = require('webpack-hot-middleware');
-var config = require('./webpack.config');
+var config = require('../webpack.config');
 
 var app = require('express')();
 var port = 3000;
